学習リスト表示のメソッドをカスタムフック化
